pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.10"}
sqlalchemy = "^2.0.38"
passlib = "^1.7.4"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
